unraid_api/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
pydantic==2.5.3
pynacl==1.5.0
pyotp==2.9.0
python-multipart==0.0.6
httpx==0.26.0
python-jose==3.3.0
passlib[bcrypt]==1.7.4
aiofiles>=23.0.0
redis>=5.0.0
websockets>=12.0
